mgr/volumes: Fail subvolume removal if it's not completed/cancelled
Removing subvolume clone which is not either complete or cancelled
with force doesn't remove the clone index (tracker). This results
in the cloner thread to stuck in loop trying to clone the deleted
one.

This patch addresses the issue by not allowing the subvolume clone
to be removed if it's not complete/cancelled even with force option.
It throws the error EAGAIN, asking the user to cancel the pending clone
and retry.

Fixes: https://tracker.ceph.com/issues/51707

+++ src/pybind/mgr/volumes/fs/operations/versions/subvolume_v2.py
@@ -148,7 +148,7 @@ class SubvolumeV2(SubvolumeV1):
                 raise VolumeException(-e.args[0], e.args[1])
         else:
             log.info("cleaning up subvolume with path: {0}".format(self.subvolname))
-            self.remove()
+            self.remove(internal_cleanup=True)
 
     def _set_incarnation_metadata(self, subvolume_type, qpath, initial_state):
         self.metadata_mgr.update_global_section(MetadataManager.GLOBAL_META_KEY_TYPE, subvolume_type.value)
@@ -337,12 +337,16 @@ class SubvolumeV2(SubvolumeV1):
         except cephfs.Error as e:
             raise VolumeException(-e.args[0], e.args[1])
 
-    def remove(self, retainsnaps=False):
+    def remove(self, retainsnaps=False, internal_cleanup=False):
         if self.list_snapshots():
             if not retainsnaps:
                 raise VolumeException(-errno.ENOTEMPTY, "subvolume '{0}' has snapshots".format(self.subvolname))
         else:
             if not self.has_pending_purges:
+                if not internal_cleanup and self.state not in [SubvolumeStates.STATE_COMPLETE, SubvolumeStates.STATE_CANCELED]:
+                    raise VolumeException(-errno.EAGAIN,
+                                          "subvolume '{0}' is a clone and not complete/cancelled."\
+                                          "Please cancel the clone and try again".format(self.subvolname))
                 self.trash_base_dir()
                 # Delete the volume meta file, if it's not already deleted
                 self.auth_mdata_mgr.delete_subvolume_metadata_file(self.group.groupname, self.subvolname)
